refactor(website): extract archive event fetch into shared useArchiveEvents hook
The archive page's inline fetch (day/tod/domain/viz-keyed, single mount fetch)
moves to shared/hooks/useArchiveEvents.ts so the upcoming installation page can
reuse the exact same fetch behavior instead of forking it. Archive page now
consumes the hook; behavior is unchanged (verified: one mount fetch, full
render).
